실행가능한 API Service 1차 완성

+++ Moyoy-Infra/src/main/java/com/moyoy/infra/github/dto/GithubProfileResponse.java
@@ -6,7 +6,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 public record GithubProfileResponse(
 
 	String login,
-	long id,
+	int id,
 
 	@JsonProperty("node_id")
 	String nodeId,
